chore: revert shared changes
