+++ crates/rome_js_analyze/src/semantic_analyzers/nursery/use_button_type.rs
@@ -1,5 +1,5 @@
 use crate::semantic_services::Semantic;
-use crate::shared::react::{is_react_create_element, PossibleCreateElement};
+use crate::utils::{is_react_create_element, PossibleCreateElement};
 use rome_analyze::{context::RuleContext, declare_rule, Rule, RuleCategory, RuleDiagnostic};
 use rome_console::codespan::Severity;
 use rome_console::markup;
