+++ crates/rome_js_analyze/src/utils/batch.rs
