+++ crates/rome_js_analyze/src/utils/rename.rs
