+++ xtask/libs_bench/benches/to_camel_case.rs
@@ -1,5 +1,5 @@
 fn to_camel_case_rome_all_lowercase() {
-    let _ = rome_js_analyze::shared::to_camel_case(iai::black_box("lowercase"));
+    let _ = rome_js_analyze::utils::to_camel_case(iai::black_box("lowercase"));
 }
 
 fn to_camel_case_case_all_lowercase() {
@@ -7,7 +7,7 @@ fn to_camel_case_case_all_lowercase() {
 }
 
 fn to_camel_case_rome_already_camel_case() {
-    let _ = rome_js_analyze::shared::to_camel_case(iai::black_box("camelCase"));
+    let _ = rome_js_analyze::utils::to_camel_case(iai::black_box("camelCase"));
 }
 
 fn to_camel_case_case_already_camel_case() {
@@ -15,7 +15,7 @@ fn to_camel_case_case_already_camel_case() {
 }
 
 fn to_camel_case_rome_pascal_case() {
-    let _ = rome_js_analyze::shared::to_camel_case(iai::black_box("CamelCase"));
+    let _ = rome_js_analyze::utils::to_camel_case(iai::black_box("CamelCase"));
 }
 
 fn to_camel_case_case_pascal_case() {
